requirements.txt
urllib3==1.26.2
gql[all]==3.0.0
autopep8==1.7.0
python-dotenv==0.21.0
pytest==7.1.3